+++ src/components/Header/Header.tsx
@@ -1,6 +1,6 @@
 import { Suspense, useEffect, useState } from 'react';
 import { useAuthState } from 'react-firebase-hooks/auth';
-import { NavLink, useNavigate } from 'react-router-dom';
+import { NavLink } from 'react-router-dom';
 
 import LanguageSelect from 'components/LanguageSelect';
 import Loader from 'components/Loader';
@@ -12,7 +12,6 @@ import styles from './Header.module.scss';
 function Header() {
   const [user] = useAuthState(auth);
   const [isSticky, setSticky] = useState(false);
-  const navigate = useNavigate();
 
   const checkScrollTop = () => {
     setSticky(window.pageYOffset > 50);
@@ -25,17 +24,6 @@ function Header() {
       window.removeEventListener('scroll', checkScrollTop);
     };
   }, []);
-
-  useEffect(() => {
-    const unsubscribe = auth.onIdTokenChanged((user) => {
-      if (!user) {
-        navigate('/');
-      }
-    });
-    return () => {
-      unsubscribe();
-    };
-  }, [navigate]);
 
   return (
     <Suspense fallback={<Loader />}>
